hiwin_pool/requirements.txt
opencv-python>=4.0
numpy>=1.20
pyserial>=3.0
